Added support for dealing with FLAGGED/UNFLAGGED
In the MS data can be flagged/unflagged. The jplotter now allows for
selecting what to plot: unflagged (default), flagged or both.
The data points will be drawn in visually distinct ways so it should be
'easy' to tell what is flagged and what isn't (if both are displayed).

Also new is better handling of NaN/Inf values - they will never be
displayed. Thus "load data/0" will produce no plots because divide-by-zero
will make everything Inf and thus not plotted.