Add support for plotting quantity versus frequency
Three new plot types 'ampfreq' 'phafreq' and 'anpfreq' were added.

For a MeasurementSet containing spectral DATA the x-axis gets transformed
from channel-within-subband to actual frequency as listed in the
SPECTRAL_WINDOW subtable.

For an MS containing LAG_DATA the x-axis gets transformed to delay in units
of seconds. The assumption is that the LAG spectrum was Nyquist sampled i.e.
the sample rate was exactly 2 * the bandwidth as per what's found in
the SPECTRAL_WINDOW subtable.