Added JIVE standardplots script
Besides attempting to implement the EVN's defined standardplots (as agreed
to be delivered with the data) it also showcases a non-interactive, yet
dynamic, interaction with "jplotter" -> it shows how you can generate
and execute jplotter commands on the fly.

This is even better than having commands in text files; this is a program
dynamically generating the commands.