"save" re-implement, extend "show", patch ppgplot
The "save" command was too simple implemented to support pdf, png outputs
(e.g. when ppgplot is linked with the Giza library). It is now
re-implemented in a more sensible way. Help file for "save" updated to
reflect this.

The "show" command (to control display of flagged/unflagged data) was
modified to also allow control of meta data display (showing header and/or
legend), now that this functionality has become available since the rewrite
of the whole plotting part (see plots.py changes).

The ppgplot module will be monkeypatched if it does not seem to have
"pgswin()"; a poor man's Python sort-of equivalent will be added. The
original "pgswin()" will be replaced by one which stores the new window for
the current device in an associative array, a "pgqwin()" is added which
looks up the window for the current device in the associative array.