allow load ppgplot from arbitray location
Sometimes there are >1 "ppgplot" Python modules on a system (e.g. parallel
system ppgplot and one linked to Giza).

Using the PYTHONPATH environment variable /does not work/ for tricking the
Python system to load ppgplot from another location.

jplotter now supports an extra command line option:
    $> jplotter ... ppgplot=/path/to/dir

to force it to use /path/to/dir/ppgplot (if that exists, of course).

Usage text updated to reflect this.

If the '-d' (debug) command line argument is given to jplotter it will
display which "ppgplot" module is currently being used.