JayB sais: "save" misses one case (ext, no type)
The "save" command needs to generate a valid PGPLOT device string:
    <file>.<ext>/<type>
and tries to be (too?) smart about leaving out any or none of <ext>, <type>,
doing some inference of either or both depending on what was given (it has
defaults if neither is given).

JayB: "save foo.ps" fails - <ext> given but <type> needs to be inferred
which it wasn't. Now it is.