Flagging for averaged data must be AND, not OR
When averaging flagged data there is a choice of how to account for flags.
Do you want to retain the knowledged that some data was flagged? then the
'averaged' mask should be logical OR of the flags. (This was
the current choice made by me)
Do you want to know if no valid data was averaged at all? Then the 'average'
mask should be the logical AND - a single, non-flagged data point would
unflag the whole averaged data. This is more what users actually want.

Thus mask 'averaging' was changed to logical_and