Add support for marking layout as fixed
jiveplot/jplotter would ALWAYS attempt to re-arrange the plot layout if it
thought it was useful to do so.

But sometimes you just want to keep the layout fixed, no matter what
jplotter was thinking. Now the "nxy" command supports marking a layout as
fixed or flexible (flexible = default because of backward compat).

See "help nxy" for details.