This fixes issue #3
JayB reported a problem when using 'animate by ****'
Turned out to be a bloody typo 'growlayout' iso 'growLayout' :-(

Also this revealed an issue with the AllInOne handling - depending on how
many plots need to be fit on the single page the layout has to be shrunk or
grown. It used to only grow, leading to unexpected results - one plot needed
to be displayed but the layout was set to 2x4 so only one of the eight
panels was drawn. Now fixed and scales it up to full page.