Add '-c <cmd>' command line option (issue #4)
It might be handy to be able to run jplotter commands from the commandline -
not unlike bash(1) 's "-c <...>" option.

This edit adds "-c <...>" support to jplotter. The option can be given
multiple times and the commands will be executed in command line order.

Example:
    $>./jplotter -c 'cd /data0/immer/ES089' -c 'ms es089_line.ms' -c r
        or, equivalent:
    $>./jplotter -c 'cd /data0/immer/ES089; ms es089_line.ms; r'

changes directory, opens MS and executes the 'r' command (range):

    ms: Current MS is 'es089_line.ms' containing 626062 rows of Spectral data for ES089 [DATA]
    listFreqs:     FREQID=1 [Doppler@G30.224]
    listFreqs:       SB 0: 6665.9620MHz/2.0MHz nch=1024 P0=RR,LL
    listAntennas:  Ef ( 0)  Hh ( 1)  Ib ( 2)  Jb ( 3)  Km ( 4)  Nt ( 5)  O8 ( 6)
    listAntennas:  T6 ( 7)  Tr ( 8)  Wb ( 9)  Ys (10)
    listSources:   J1834-0301 G30.224 G30.198 3C345
    listTimeRange: 04-Mar-2018/04:00:01.000 -> 04-Mar-2018/11:59:58.000  dT: 2.000