fix multi=t when loading computed data sets
With multi=true the *-versus-channel plotter tries to put the band edge in
the plot, for which it must load the channel->frequency mapping from the
'msname' attribute of the plots.

This attribute, if a data set expression is used
(e.g. "load stored_data_set1 - stored_data_set2"), gets set to the
expression and not the file-on-disk.

This failure was ignored silently but later on code would try to use the
(non-existant) spectral mapping unconditionally. Now it is used only if
there IS a spectral mapping.