ckey support for attr[none] and default = <colour>
Previously, with ckey special colouring assignments, it was very easy to get
an error "no match for XXXX" if e.g. the attribute you selected on was
stripped out.

Example:
    > ckey p[ll]=2 p[rr]=3

This could fail in two ways:
    1 plotting single pol data
    2 plotting data containing other pols besides ll,rr

1) in this case, all data sets have the same value for the P attribute and
hence it gets stripped and set to None. "None" does not match with either
"ll" or "rr" resulting in the no-match error

2) in this case, there are data sets that have a value in their P attribute
that is neither "ll" or "rr", also resulting in the no-match error.

The grammar and implementation of ckey has been extended:

1) "ckey ... <attribute>[none]=2" is now possible, you can match on an explicit "none"
for the attribute value

2) "ckey .... default = <colour>" is now possible: a default if all
preceding matches have failed.

Both can be combined and are documented in the help for ckey.