Add ".extend()" to dataset_list for *-vs-uvdist
because uv distance scales with baseline length and channel frequency. So
we can compute a number of uv distances in one go (for each channel) and add
all of the results as x-axis in one go. Way more efficient than iterating
over each channel and adding them point by point