Fix issue #8 labels appear to be written multiple
Turns out they weren't painted twice. The linewidth was set to "4" when
drawing PGTXT() and to screen that made no difference to when plotting to
file PGPLOT suddenly starts to honour that attribute and draw the text in
(very) bold characters.
Is now fixed by alternating line width between line and text to an
appropriate setting.