Singularity image now tagged ":latest", security changes
The jiveplot docker image is now always tagged so the Singularity image should follow that too.

Also there has been a Singularity hub security policy change - it is advised to not run the container from the "shub://..." URL directly anymore. The amount of API calls to the "shub://" repository is now *limited*.

See the README.md for updated details