new approach: recognize antenna names!
Brainwave: define all the antenna names as tokens in the selection parser!
Now a baseline typed as "efwb" will be automatically parsed as two antennas
(if they exist in the dataset)!

This makes the parser simpler.

The TaQL can now be generated directly after actually parsing the baseline
specification - the antenna id can be looked up or it is given (integer).

So the TaQL is shorter and typically faster.

Also allows the code to warn if you specify an antenna that is not present
in the data