Add comparison postprocessing step
For comparing two datasets for equal data content the
    postprocessing/compare_data.compare_data()

function can be used. If loaded, it always stores a copy of the latest plot that is generated after loading.

Each time a plot is made, and its plot type is the same as the currently stored data set, all data sets between the new and stored data are compared; all differences between x- and y- arrays for data sets common to both plots are accumulated for both x and y.

If any of the accumulated arrays (x and/or y) contain a difference > set tolerance (currently 1e-7), the compare_data() will raise a RuntimeError.

If the new and previous plots do not have exactly the same data sets or have differing plot types, they compare unequal w/o even trying and a RuntimeError is raised.

After the comparison but before raising a runtime error the new data set is installed as the new "last" dataset.

Suggested sequence for testing:

    # install this module
    > postprocess postprocessing/compare_data.compare_data

    # Define which information and subset to extract that is expected to be numerically equivalent
    # between the two data sets. Make sure the 'pl' command is last - it triggers extraction of
    # the data and generates the plot datasets
    > macro extract "pt rnichan; fq ...; bl ...; indexr; scan nnn; pl"

    # Open reference data set and extract the relevant info
    > ms reference.ms;
    > extract

    # Now open data set that is expected to have
    # the exact same numerical content, by
    > ms test.ms
    > extract

Profit!