Finally (wow ...) stop execution on first error ...
Previously, if entering a sequence of commands:

    jcli> pt amptime; ch mid; avt none; scan 17; bl cross -mc; pl

and e.g. the `indexr` command wasn't run then the "scan 17" command would
fail with an error message but the interpreter would continue on executing
commands *anyway* ... this is really not good!

Now it stops executing commands on error ... only after a decade or so ;-)