Fix unpacking error leftovers
When going from Py2 => Py3 needed to fix a lot of "unpacking tuples in
function arguments", which useful functionality was removed in Py3.

    def f( (x,y) ):
        print("x=", x, "y=", y)

needed to be rewrittten to:

    def f( x_y ):
        print("x=", x_y[0], "y=", x_y[1])

and there were a few instances left over where this process wasn't done
correctly, but also this was in code that wasn't actually used by the
program, apparently.