Fix bug in baseline parser on "bl none"
In order to make the baseline parser more logical such that:
    bl none +jb*
works as expected ("remove selection, then add baselines to jb") [note:
which is equivalent to "bl jb*" ...) the "none" selector would yield
TaQL string "FALSE".

This string made it into the final TaQL such that a simple:
    bl none
(to clear any previous selection) would add "FALSE AND" to the total
selection string - resulting in not a single row of data being selected.

Now fixed that if, at the end of the parsing of the baseline selection
command, only 'FALSE' remains, it is removed.

Also needed to add some extra parentheses to make sure the logic gets
evaluated correctly *cough*