Handle ^C slightly different; do NOT ALWAYS redraw
Plotting could be interrupted by pressing ^C (e.g. if you realised it would take
a loooong time b/c of an erroneous selection).

However, after ^C, the screen would be redrawn always - and if it was a huge
plot could still take a lot of time.

Now the ^C is handled at higher level such that the drawing of that:
- can be interrupted too
- the redrawring of the screen is now conditional if the data selecting
  changed

Currently that means that if any of the plotting *attributes* (marking, line
width, symbol, layout - you name it) changes, the plot will not be
refreshed.

Temp solution: add "--force" option to the "pl" command to force redraw.

Next commit will be detecting changes in plot attributes.