fix erroneously raised unknown DATA_DESC_ID  error
The code was looking in all frequency groups and subbands defined therein to
find a given DATA_DESC_ID but the handling of the returned result(s) was a
bit too aggressive:

if the requested DATA_DESC_ID was not found in the first result the code
would raise an "Invalid DATA_DESC_ID" without checking other results first.

This error should only be raised if all the returned results have been
exhausted w/o finding a match (which it does now).

There is one other intermediate check that currently throws an exception: in
case > 1 spectral windows claim to be linked to the same DATA_DESC_IDs. In
that case it is rather impossible to uniquely unmap a DATA_DESC_ID into a
spectral window.