remove debug print, add useful warnings
When combining data sets using "load X + Y" or "store 2*X / Y" where both X
and Y are sets of plots there are several possible (unexpected) outcomes and the s/w
would not inform you about *any* of them or whether they'd occurred. This is
now fixed.

Background: the syntax "X <operator> Y" triggers a combining function which
will apply the operator on the datasets in X and Y that have the same key -
in pseudo-code:

    for key in intersect(X.keys(), Y.keys()):
        result[key] = operator(X[key], Y[key])

Now, if X and Y do not have the same number of keys, or the keys differ
between them, the number of keys in the result will be either 0 (no common
keys between X- and Y) or not equal to the number of keys in X or Y.

The software didn't use to check and just give you back "result", the
content depending on the actual intersection of the sets of keys in X and
those in Y and not warn, ever.

Now the code checks & warns for the different cases and issues an
appropriate warning