It's 2025. pyproject.toml + PEP328 compliance
Big one. Time to get to grips with the "future".
Python packaging has changed a lot and it's long overdue to reorganize
the project into a proper python package.

There is now an installable python module "jiveplot" with a
pyproject.toml and setup.cfg.

Now can use:

    $> pip install -e .

if you want to install it from a git clone.

Also addressed PEP328:
    - ok I give in: one (system) import per line (jeez, but ok)
    - all imports loading multiple items have been rewritten to use "(...)" grouping
    - use relative imports for own modules using leading dot

One thing that does not work anymore is loading a postprocessing
function that came with the package, need to work out how to make that
work again.

Can always specify absolute path to "/path/to/some/file.function" to
load "function" from "file.py" as post-processing method but that's a
bitch. Want to be able to ship maintained post-processing functions.