Fix <map object> and "map_" mess-up
Found that when using source selection the TaQL would (again) be broken:

    Error in TaQL command: using style Python select  from $1 where
            ((((ANTENNA1==11 AND ANTENNA2==2)) OR ((ANTENNA1==2 AND ANTENNA2==11)))) AND
            (FIELD_ID IN <map object at 0x104fd0220>)
    parse error at or near position 157 '0x104fd0220'

Turns out there was a vanilla "map(...)" still hiding in there, which
means that in Py3 that's not a list but a map object ...

Also noticed that in the same source selection function there was a
statement:

    map_ = self.selection.mapping # or words to that effect

but there'se also:

    from .functional import (..., map_, ...)

so we were overwriting our own "map_" variable. Yikes.
I really would like a compiled language for this tbh