Configure correct(new) dependency name
The Python PGPLOT extension had to be renamed since the attempt to
publish on PyPI. There was an already existing package on there with the
name "ppgplot" but, however, was completely unrelated.

So the PyPI project name was changed to "python-pgplot", so we need to
refer to that as dependency.